requirements.txt
Django==2.1.7
# Editable install with no version control (django-rest-hydra==0.1.0)
-e /home/invinciblycool/Documents/Github-projects/django-rest-hydra
djangorestframework==3.9.2
pytz==2018.9
git+https://github.com/HTTP-APIs/hydra-python-core@v0.1#egg=hydra_python_core
git+https://github.com/HTTP-APIs/hydra-openapi-parser@0.1.1#egg=hydra_openapi_parser
